% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{CAC40}
\alias{CAC40}
\title{Paris stock exchange}
\format{
A vector with the variable \code{Close}.

 There are 7936 observations. We removed every \code{NULL} values.
}
\source{
Data pulled from Yahoo Finance: \samp{https://fr.finance.yahoo.com/quote/\%5EFCHI/history?p=\%5EFCHI}
}
\usage{
CAC40
}
\description{
This data set considers market index at the closure of the market from
March 1, 1990 to June 14, 2021.
}
\seealso{
\code{\link{CAC40return}} and \code{\link{CAC40return.sq}}
}
\keyword{datasets}
