% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{meansq}
\alias{meansq}
\title{Function optim will minimize}
\usage{
meansq(x, dim.ar = NULL, dim.ma = NULL, y)
}
\arguments{
\item{x}{One point in \eqn{\rm I\!R^{(p+q)}}.}

\item{dim.ar}{Length of AR vector.}

\item{dim.ma}{Length of MA vector.}

\item{y}{Vector of a time series.}
}
\value{
\describe{
   \item{\code{ms}}{Mean square at the point \code{x}.}
}
}
\description{
Computes the mean square of the time series at the point \code{x}, will be minimize with the \code{\link[stats]{optim}} function in our function \code{\link{estimation}}.
}
