% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VARest.R
\name{VARest}
\alias{VARest}
\title{Estimation of VAR(p) model}
\usage{
VARest(x, p)
}
\arguments{
\item{x}{Matrix of dimension (n,p+q).}

\item{p}{Integer for the lag order.}
}
\value{
A list containing:
\describe{
   \item{\code{ac}}{Coefficients data matrix.}
   \item{\code{p}}{Integer of the lag order.}
   \item{\code{k}}{Dimension of the VAR.}
   \item{\code{res}}{Matrix of residuals.}
}
}
\description{
Estimates the coefficients of a VAR(p) model. Used in \code{\link{matXi}}.
}
